requirements.txt
gensim>=4.3.0
tokenizers>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
requests>=2.31.0
umap-learn>=0.5.5
tqdm>=4.66.0
